refactor search facility mock responses to be clearer
