+++ src/applications/caregivers/tests/mocks/fetchFacility/index.js
@@ -1,0 +1,2 @@
+export * from './mockVetsApiResponses';
+export * from './mockFormattedResponses';
